+++ docs/conf.py
@@ -111,7 +111,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'pyocrddoc'
+htmlhelp_basename = 'coredoc'
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -138,7 +138,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'pyocrd.tex', u'OCR-D/core Documentation',
+    (master_doc, 'core.tex', u'OCR-D/core Documentation',
      u'Kay Wuerzner, Konstantin Baierer', 'manual'),
 ]
 
@@ -148,7 +148,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'pyocrd', u'OCR-D/core Documentation',
+    (master_doc, 'core', u'OCR-D/core Documentation',
      [author], 1)
 ]
 
@@ -159,8 +159,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'pyocrd', u'OCR-D/core Documentation',
-     author, 'pyocrd', 'One line description of project.',
+    (master_doc, 'core', u'OCR-D/core Documentation',
+     author, 'core', 'One line description of project.',
      'Miscellaneous'),
 ]