+++ ocrd_models/ocrd_models/ocrd_file.py
@@ -32,6 +32,13 @@ class OcrdFile():
         self._instance = instance
         self.mets = mets
 
+        if self.url:
+            if not self.url.startswith('http://') and \
+                not self.url.startswith('https://'):
+                self.local_filename = self.url
+            if self.url.startswith('file://'):
+                self.local_filename = self.url[len('file://'):]
+
         #  if baseurl and not local_filename and '://' not in self.url:
         #      self.local_filename = '%s/%s' % (baseurl, self.url)
 
@@ -56,6 +63,13 @@ class OcrdFile():
         Get the ``os.path.basename`` of the local file, if any.
         """
         return os.path.basename(self.local_filename)
+
+    @property
+    def extension(self):
+        basename, ext = os.path.splitext(self.basename)
+        if basename.endswith('.tar'):
+            ext = ".tar" + ext
+        return ext
 
     @property
     def basename_without_extension(self):