+++ ocrd_utils/ocrd_utils/logging.py
@@ -101,7 +101,7 @@ def initLogging():
     for p in CONFIG_PATHS:
         config_file = os.path.join(p, 'ocrd_logging.py')
         if os.path.exists(config_file):
-            logging.info("Loading logging configuration from '%s'", config_file)
+            logging.debug("Loading logging configuration from '%s'", config_file)
             with open(config_file) as f:
                 code = compile(f.read(), config_file, 'exec')
                 exec(code, globals(), locals()) # pylint: disable=exec-used